Implemented HTTP Basic Authentication for routes
