+++ index.js
@@ -1,5 +1,7 @@
 const express = require('express');
 const mongoose = require('mongoose');
+const passport = require('passport');
+const { BasicStrategy } = require('passport-http');
 const bodyParser = require('body-parser');
 const path = require('path');
 
@@ -13,14 +15,26 @@ require('./models/ClassSlot');
 
 mongoose.connect(keys.mongoURI);
 
-// Create express server
+// Initialize Passport Authentication
+passport.use(
+  new BasicStrategy((username, password, cb) => {
+    if (username === keys.authUsername && password === keys.authPassword) {
+      return cb(null, true);
+    }
+    return cb(null, false);
+  })
+);
+
+// Create Express server
 const app = express();
 
 // Define Middleware
 app.use(bodyParser.json());
+app.use(passport.initialize());
+app.use(passport.session());
 
 // Import routes
-require('./routes/scrapeRoutes')(app);
+require('./routes/authenticatedRoutes')(app);
 require('./routes/apiRoutes')(app);
 
 if (process.env.NODE_ENV === 'production') {
